add holiday replacement automatically after selection
this commit introduces thymeleaf as template engine.
I think this is the time we can move to an up to date and faster template engine :-)

I don't know how to render partials with JSP
So the template code can also be copied with thymeleaf.

Let's see how the template engine transition works ¯\_(ツ)_/¯